Unit tests2

+++ src/main/java/com/cloudDomus/cloudDommus/Client/ClientController.java
@@ -5,6 +5,7 @@ import java.util.List;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 
 @CrossOrigin(origins = "http://localhost:8080", maxAge = 3600)
@@ -15,6 +16,10 @@ public class ClientController {
 
     @Autowired
     ClientRepository repository;
+
+    public void setRepository(ClientRepository repository) {
+        this.repository = repository;
+    }
 
     // Aggregate root
     @ApiOperation(value = "View a list of available Client", response = List.class)
@@ -34,13 +39,20 @@ public class ClientController {
     @GetMapping("/client/{id}")
     public Client getClientByID(@PathVariable Long id) {
 
-        return repository.findById(id).orElseThrow(() -> new ClientNotFoundException(id));
+        return repository.findByID(id);
     }
 
     @ApiOperation(value = "Delete a work by its Id", response = List.class)
     @DeleteMapping("/Client/{id}")
-    public void deleteClient(@PathVariable Long id) {
+    @Transactional(rollbackFor = ClientNotFoundException.class)
+    public Client deleteClient(@PathVariable Long id)throws ClientNotFoundException{
+        Client deleted = repository.deleteByID(id);
+        if(deleted == null){
+            throw new ClientNotFoundException(id);
+        }
         repository.deleteById(id);
+        deleted.setId(null);
+        return deleted;
     }
 
 }
